refactor: 1. 메소드 명 및 반환 String 조정 2. CommentNotFoundException, UserNotFoundException 추가
