+++ src/main/java/project/como/domain/comment/exception/CommentNotFoundException.java
@@ -1,0 +1,11 @@
+package project.como.domain.comment.exception;
+
+import project.como.global.common.exception.ComoException;
+import project.como.global.common.exception.ErrorType;
+
+public class CommentNotFoundException  extends ComoException.NotFoundException {
+
+	public CommentNotFoundException(Long commentId) {
+		super(ErrorType.NotFound.COMMENT_NOT_FOUND, "ID가 " + commentId + "인 게시글을 찾을 수 없습니다.");
+	}
+}
