+++ src/main/java/project/como/domain/comment/repository/CommentRepository.java
@@ -1,4 +1,4 @@
-package project.como.domain.comment.reposiory;
+package project.como.domain.comment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
